feat(routes) Arquivos do front atualizados
